Specify prototypes for used X11 C functions
Python does not magically resolve types of C functions imported through
ctypes. Without specifying their prototypes the return type of all
functions is c_int and all Python int arguments are converted to C int
arguments.

The return type of XOpenDisplay is 'Display *', so the correct return type is
'c_void_p'.

The argument type of XCloseDisplay is 'Display *', so the correct
argument type is 'c_void_p' too.

Without this patch, the pointers returned from XOpenDisplay and passed
to XCloseDisplay are truncated to int, which is shorter than 'void *' on
some architectures, and that causes unpredicted behavior which leads to
a crash.

Related: bugzilla.redhat.com/1244261

Signed-off-by: Jakub Filak <jfilak@redhat.com>